Implement initial support for generic Move structs. (#171)
This patch implements support for generic structs (structs parameterized by types). This builds on the previous patch that implements concrete struct support. That code eagerly detects and declares structs before the rest of the translation. We currently continue the same approach for generics.

The idea is to first visit the various "instantiation" tables (e.g., StructDefInstantiations, FieldInstantiations) to locate and collect the generic types.

Each generic type is then expanded into multiple concrete LLVM struct types depending on usage in the module.

In some cases, not all generic types are evident from the tables mentioned above. These can arise implicitly from usage as function arguments or return values, or when constructing a generic. These remaining uses are gleaned from the signature table and added to the list to be expanded.

Note that llvm_type is currently not general enough to handle generics and does not properly deal with TypeParameter. I've written a wrapper type getter that is general enough to handle many cases. This avoids needing to update all the users of llvm_type in this patch (i.e., additional unnecessary churn). An orthogonal clean-up task is to generalize the existing llvm_type and is left for a later patch.

A few runnable rbpf test cases are included as well as a couple of static LLVM IR tests. As time goes on, we'll feed more substantial code through and add additional tests.

Finally, it is worth mentioning that in the original concrete struct patch, I considered lazily declaring structs during the translation proper. At the time, that was more complex than the eager approach (i.e., doing it ahead of time). With generics, however, the lazy approach might be simpler (i.e., see the passage above about cases not directly seen in the instantiation tables). I may re-examine this later as time permits.